fix: unable to build via buildx
While I'm not 100% sure _why_ this is the case, something in Erlang's
make system blows up when building the image via buildx. Enabling the
`--allow-insecure-entitlement security.insecure` on the buildx instance
fixes the issue, so I'm assuming it is something to do with how the
build system interacts with the operating system, but I can't get enough
trace data to understand what exactly, and it isn't reproducible outside
of buildx, so troubleshooting steps have to be performed as part of the
Dockerfile RUN commands.

Furthermore, even with this fix, there appears to be some kind of build
issue/race with the `snmp` application, which fails almost every time
due to a missing target, but sometimes it succeeds, and I can't figure
out what the precondition is. Presumably this will be fixed as part of
the ongoing changes to the build system, but for now I've disabled the
snmp application.